move DiscoveryNamespaceClient to data fabric
