+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/util/DiscoveryNamespaceClient.java
@@ -14,11 +14,12 @@
  * the License.
  */
 
-package co.cask.cdap.common.namespace;
+package co.cask.cdap.data2.util;
 
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.discovery.EndpointStrategy;
 import co.cask.cdap.common.discovery.RandomEndpointStrategy;
+import co.cask.cdap.common.namespace.AbstractNamespaceClient;
 import co.cask.common.http.HttpRequest;
 import co.cask.common.http.HttpRequests;
 import co.cask.common.http.HttpResponse;
